+++ src/hoc/withTheme.js
@@ -14,12 +14,13 @@ import determineTheme from '../utils/determineTheme'
 
 const wrapWithTheme = (Component: ReactClass<any>) => {
   const componentName = Component.displayName || Component.name || 'Component'
+  const isStatelessFunctionalComponent =
+    typeof Component === 'function' &&
+    !(Component.prototype && 'isReactComponent' in Component.prototype)
 
+  // NOTE: We can't pass a ref to a stateless functional component
   const shouldSetInnerRef =
-    _isStyledComponent(Component) ||
-    // NOTE: We can't pass a ref to a stateless functional component
-    (typeof Component === 'function' &&
-      !(Component.prototype && 'isReactComponent' in Component.prototype))
+    _isStyledComponent(Component) || isStatelessFunctionalComponent
 
   class WithTheme extends React.Component {
     static displayName = `WithTheme(${componentName})`
@@ -78,18 +79,17 @@ const wrapWithTheme = (Component: ReactClass<any>) => {
     }
 
     render() {
-      // eslint-disable-next-line react/prop-types
-      const { innerRef } = this.props
-      const { theme } = this.state
+      const props = {
+        theme: this.state.theme,
+        ...this.props,
+      }
 
-      return (
-        <Component
-          theme={theme}
-          {...this.props}
-          innerRef={shouldSetInnerRef ? innerRef : undefined}
-          ref={shouldSetInnerRef ? undefined : innerRef}
-        />
-      )
+      if (!shouldSetInnerRef) {
+        props.ref = props.innerRef
+        delete props.innerRef
+      }
+
+      return <Component {...props} />
     }
   }